Refactor and modularize character and tag services
Extracted character and tag-related logic into dedicated service files (`characters-service.js`, `tags-service.js`) to improve modularity and maintainability. This includes creating utility functions for managing characters, tags, and related operations. Updated relevant imports and removed redundant code to align with the new structure.
