+++ src/components/tags.js
@@ -1,25 +1,9 @@
-import { equalsIgnoreCaseAndAccents, includesIgnoreCaseAndAccents } from '../utils.js';
 import { tags } from "../../../../../tags.js";
 import { tagFilterstates } from "../constants/settings.js";
-import {
-    tagList,
-    tagMap,
-    power_user,
-    saveSettingsDebounced
-} from "../constants/context.js";
-import {addTagToCategory} from "../services/settings-service.js";
-import {refreshCharList} from "./characters.js";
-
-export { displayTag, generateTagFilter, tagFilterClick, addListenersTagFilter, renameTagKey, createTagInputCat };
-
-/**
- * Initializes tag filter states
- */
-export function initializeTagFilterStates() {
-    tagList.forEach(tag => {
-        tagFilterstates.set(tag.id, 1);
-    });
-}
+import { tagList } from "../constants/context.js";
+import { addTagToCategory } from "../services/settings-service.js";
+import { refreshCharListDebounced } from "./characters.js";
+import { findTag } from "../services/tags-service.js";
 
 /**
  * Renders a tag as an HTML string based on the provided tag ID and an optional category flag.
@@ -28,7 +12,7 @@ export function initializeTagFilterStates() {
  * @param {boolean} [isFromCat=false] - Indicates whether the tag is from a category.
  * @return {string} The HTML string representation of the tag. Returns an empty string if the tag ID is not found in the tag list.
  */
-function displayTag( tagId, isFromCat = false ){
+export function displayTag( tagId, isFromCat = false ){
     const tagClass = isFromCat ? "fa-solid fa-circle-xmark tag_cat_remove" : "fa-solid fa-circle-xmark tag_remove";
     if (tagList.find(tagList => tagList.id === tagId)) {
         const name = tagList.find(tagList => tagList.id === tagId).name;
@@ -52,12 +36,12 @@ function displayTag( tagId, isFromCat = false ){
 }
 
 /**
- * Generates and displays an HTML block of sorted tags with specific styles and attributes,
+ * Generates and displays an HTML block of sorted tags with specific styles and attributes
  * and initializes filter states for these tags.
  *
  * @return {void} This function does not return a value.
  */
-function generateTagFilter() {
+export function generateTagFilter() {
     let tagBlock='';
 
     tagList.sort((a, b) => a.name.localeCompare(b.name));
@@ -69,20 +53,6 @@ function generateTagFilter() {
     });
 
     $('#tags-list').html(tagBlock);
-}
-
-/**
- * Adds event listeners to elements with the class 'acm_tag' to handle click events.
- * Each click event triggers the `tagFilterClick` function with the corresponding tag element.
- *
- * @return {void} This method does not return a value.
- */
-function addListenersTagFilter() {
-    const tags = document.querySelectorAll('.acm_tag');
-
-    tags.forEach(tag => {
-        tag.addEventListener('click', () => tagFilterClick(tag));
-    });
 }
 
 /**
@@ -100,7 +70,7 @@ function addListenersTagFilter() {
  *
  * @return {void} This function does not return a value.
  */
-function tagFilterClick(tag) {
+export function tagFilterClick(tag) {
     const currentState = tagFilterstates.get(tag.id);
     let newState;
 
@@ -120,44 +90,7 @@ function tagFilterClick(tag) {
     }
 
     tagFilterstates.set(tag.id, newState);
-    refreshCharList();
-}
-
-/**
- * Renames a tag key in the tag map by transferring the corresponding value to a new key
- * and removing the old key from the tag map.
- *
- * @param {string} oldKey - The existing tag key to be renamed.
- * @param {string} newKey - The new name for the tag key.
- * @return {object} tag - Returns the updated tag map after the rename operation.
- */
-function renameTagKey(oldKey, newKey) {
-    const value = tagMap[oldKey];
-    tagMap[newKey] = value || [];
-    delete tagMap[oldKey];
-    saveSettingsDebounced();
-}
-
-/**
- * Finds tags based on the provided request, resolving the result with filtered and sorted tags that match the search term.
- *
- * @param {Object} request - The search request containing a `term` property to match tags.
- * @param {Function} resolve - A callback function to resolve the results array.
- * @param {string} listSelector - Selector for the list element containing tags, used to exclude tags already present in the list.
- * @return {Array<string>} - The filtered and sorted list of tag names matching the search term, including the term itself if no exact match is found.
- */
-function findTag(request, resolve, listSelector) {
-    const skipIds = [...($(listSelector).find('.tag').map((_, el) => $(el).attr('id')))];
-    const haystack = tagList.filter(t => !skipIds.includes(t.id)).sort(compareTagsForSort).map(t => t.name);
-    const needle = request.term;
-    const hasExactMatch = haystack.findIndex(x => equalsIgnoreCaseAndAccents(x, needle)) !== -1;
-    const result = haystack.filter(x => includesIgnoreCaseAndAccents(x, needle));
-
-    if (request.term && !hasExactMatch) {
-        result.unshift(request.term);
-    }
-
-    resolve(result);
+    refreshCharListDebounced();
 }
 
 /**
@@ -167,7 +100,7 @@ function findTag(request, resolve, listSelector) {
  * @param {string} listSelector - the selector for the list of the tags modified by the input control
  * @param {PrintTagListOptions} [tagListOptions] - Optional parameters for printing the tag list. Can be set to be consistent with the expected behavior of tags in the list that was defined before.
  */
-function createTagInputCat(inputSelector, listSelector, tagListOptions = {}) {
+export function createTagInputCat(inputSelector, listSelector, tagListOptions = {}) {
     $(inputSelector)
         // @ts-ignore
         .autocomplete({
@@ -175,31 +108,7 @@ function createTagInputCat(inputSelector, listSelector, tagListOptions = {}) {
             select: (e, u) => selectCatTag(e, u, listSelector, { tagListOptions: tagListOptions }),
             minLength: 0,
         })
-        .focus(onTagInputFocus); // <== show tag list on click
-}
-
-/**
- * Compares two given tags and returns the compare result
- *
- * @param {Tag} a - First tag
- * @param {Tag} b - Second tag
- * @returns {number} The compare result
- */
-function compareTagsForSort(a, b) {
-    const defaultSort = a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-    if (power_user.auto_sort_tags) {
-        return defaultSort;
-    }
-
-    if (a.sort_order !== undefined && b.sort_order !== undefined) {
-        return a.sort_order - b.sort_order;
-    } else if (a.sort_order !== undefined) {
-        return -1;
-    } else if (b.sort_order !== undefined) {
-        return 1;
-    } else {
-        return defaultSort;
-    }
+        .focus(onTagInputFocus); // <== show a tag list on click
 }
 
 /**
