+++ src/events/characterDetails-events.js
@@ -1,11 +1,10 @@
-import { editCharDebounced } from "../components/characters.js";
 import { selectedChar, setMem_avatar } from "../constants/settings.js";
 import { getTokenCount, selectCharacterById, substituteParams } from "../constants/context.js";
 import { getIdByAvatar } from "../utils.js";
 import { addAltGreeting, delAltGreeting } from "../components/altGreetings.js";
 import { setCharacterId } from '../../../../../../script.js';
 import { closeDetails } from "../components/modal.js";
-
+import { editCharDebounced } from "../services/characters-service.js";
 
 
 export function initializeCharDetailsEvents() {
