+++ src/events/tags-events.js
@@ -1,0 +1,13 @@
+/**
+ * Adds event listeners to elements with the class 'acm_tag' to handle click events.
+ * Each click event triggers the `tagFilterClick` function with the corresponding tag element.
+ *
+ * @return {void} This method does not return a value.
+ */
+export function addListenersTagFilter() {
+    const tags = document.querySelectorAll('.acm_tag');
+
+    tags.forEach(tag => {
+        tag.addEventListener('click', () => tagFilterClick(tag));
+    });
+}
